libs/common-python/pyproject.toml
[project]
name = "common-python"
version = "0.1.0"
description = "Shared chassis library for TMF Product Catalog microservices"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-json-logger>=2.0.7",
    "python-jose[cryptography]>=3.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "aio-pika>=9.3.1",
    "asyncpg>=0.29.0",
    "camunda-external-task-client-python3>=0.1.0",
    # OpenTelemetry Core
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    # Exporters
    "opentelemetry-exporter-zipkin>=1.21.0",
    # Instrumentation
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-httpx>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-instrumentation-aio-pika>=0.42b0",
    # Propagators
    "opentelemetry-propagator-b3>=1.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
